requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
Brotli==1.1.0
bs4==0.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
cmdstanpy==1.2.4
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.3.0
Django==5.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
holidays==0.61
html5lib==1.1
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.3
ipython==8.22.1
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.18
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.3
jupyter-nbextensions-configurator==0.6.3
jupyter-tabnine==1.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.10
jupyterthemes==0.20.0
keras==3.6.0
kiwisolver==1.4.5
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-text-splitters==0.3.2
langsmith==0.1.147
lazy_loader==0.4
lesscpy==0.15.1
libclang==18.1.1
livekit-api==0.6.0
livekit-protocol==0.6.0
llvmlite==0.43.0
lxml==5.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.3
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.3
patsy==1.0.1
peewee==3.17.8
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
prometheus_client==0.20.0
prompt-toolkit==3.0.43
prophet==1.1.6
protobuf==5.27.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.17.2
PyJWT==2.9.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.17
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
Quandl==3.7.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.0
safetensors==0.4.5
scikit-image==0.23.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==3.3.1
setuptools==69.0.2
sitemap==20191121
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
stanio==0.5.1
starlette==0.27.0
statsmodels==0.14.4
stumpy==1.13.0
sympy==1.13.1
ta==0.11.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
terminado==0.18.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2024.2.12
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.0.1
torch==2.5.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.46.3
tsfresh==0.20.3
types-protobuf==4.25.0.20240417
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.24.0.post1
virtualenv==20.25.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
wheel==0.42.0
whitenoise==6.8.2
widgetsnbextension==4.0.10
wordcloud==1.9.3
wrapt==1.16.0
yapf==0.40.2
yarl==1.9.4
yfinance==0.2.50
zipp==3.17.0
